Refactor Interface to Interfaces Interface is a reserved word for namespacing
