+++ tripal/src/TripalVocabTerms/PluginManagers/TripalVocabularyManager.php
@@ -32,7 +32,7 @@ class TripalVocabularyManager extends TripalCollectionPluginManager {
         ,$namespaces
         ,$cache_backend
         ,$module_handler
-        ,'Drupal\tripal\TripalVocabTerms\Interface\TripalVocabularyInterface'
+        ,'Drupal\tripal\TripalVocabTerms\Interfaces\TripalVocabularyInterface'
         ,'Drupal\tripal\TripalVocabTerms\Annotation\TripalVocabulary'
         ,'tripal_vocabulary_collection'
     );
